backend/requirements.txt
asgiref==3.8.1
Django==v4.2.16
djangorestframework==3.15.2
sqlparse==0.5.1
typing_extensions==4.12.2
pyjwt
Pillow
djangorestframework-simplejwt
psycopg
psycopg2
pydevd-pycharm
django-cors-headers==3.13.0
django-filter